Refactoring to allow BoundLogger usage at init
As noted in #469, `make_filtering_bound_logger` was returning
a `BoundLoggerBase` with minimal functionality. These changes
refactor the `BoundLogger` definition to avoid circular dependencies
and allow for it's use in `make_filtering_bound_logger`.
